feat(api): export target-resolver builder as TargetResolverConfig.BuildResolver
Move the target-resolver builders out of the serve command into pkg/api and
expose them as a single method, TargetResolverConfig.BuildResolver, so embedders
can build the same Etre / Secrets-Manager resolver from rendered config instead
of reaching into serve internals.

Pure move + export — no behavior change. serve.go now calls
config.TargetResolver.BuildResolver(ctx, logger); the builder helpers and their
tests move to pkg/api alongside it.
